requirements.txt
alembic==1.5.7
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
numpy==1.20.1
pandas==1.2.3
protobuf==3.15.6
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
stanza==1.2
torch==1.8.0
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.4
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
